requirements.txt
altgraph==0.17.3
cx-Freeze==6.15.5
cx-Logging==3.1.0
et-xmlfile==1.1.0
lief==0.13.2
numpy==1.24.3
openpyxl==3.1.0
pandas==1.5.3
pefile==2023.2.7
Pillow==10.0.0
psutil==5.9.5
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.2
scipy==1.11.1
six==1.16.0
tzdata==2023.3
xgboost==1.7.6
XlsxWriter==3.1.0
